Remove "archlinux" and "blackarch" from pacman-key --populate
At time of line 193, blackarch.gpg file does not exist, so by running `pacman-key --populate archlinux blackarch` generates the error:
```
==> ERROR: The keyring file /usr/share/pacman/keyrings/blackarch.gpg does not exist.
```
If strap.sh is run in an interactive shell by the user, this error is not blocking. If the script is run by another script inside a process (for example fakechroot during Make), the error will interrupt the process.

The cleanest solution is to replace the line as proposed.